Moving admin stats controller to a behavior so it can be overriden in the main app
